refactor: move servemux to gphttp/servemux
